Output hosts table as xml or json.
Originally tried to do this with django's built-in serialization, but
found it a bit too limited. Our tables have multi-column primary keys,
which django doesn't know how to handle, and which confused the
serializers. Working around this limitation for the json serializer was
doable (by subclassing it and overriding get_dump_object to to force the
natural_key to be used) but the xml serializer was more complicated and
it seemed easier just to write a few lines of django template to do it.